Initialize "fur_" before "random_waves_".
The C++ standard states that a class member's constructor is
initialized in the order in which it appears in the class definition:
http://stackoverflow.com/questions/2669888/c-initialization-order-for-member-classes

Before this change, however, RandomWaveVisualizer::RandomWaveVisualizer(Fur*)
assumed that fur_ was already initialized and accessed the member.  Since fur_
was uninitialized, the behavior was undefined and on g++ 4.8.2 (Ubuntu 14.02).
The allocation loop executed (effectively) infinately until memory was
exhausted and a SIGABRT was raised.  It is unclear what happened to make this
behavior not exhibit itself in other environments.

This change moves the fur_ declaration in the class definition so that
initialization occurs in the order that the rest of the code assumes.  This
fixes the problem.

Stacktrace of the fault that occurred before this error:

Program received signal SIGABRT, Aborted.
0x00007ffff6011cc9 in __GI_raise (sig=sig@entry=6) at ../nptl/sysdeps/unix/sysv/linux/raise.c:56
56      ../nptl/sysdeps/unix/sysv/linux/raise.c: No such file or directory.
(gdb) where